Extract PlatformDetection to a common test helper
Although PlatformDetection in System.Globalization.Extensions is a
subset of PlatformDetection in System.Globalization, there's still
duplicated code there. Unfortunately, a dependency on
System.IO.FileSystem has to be added to System.Globalization.Extensions,
but this is a test project, so isn't such a big deal
