+++ src/System.Globalization.Extensions/tests/project.json
@@ -4,6 +4,7 @@
     "System.Console": "4.0.0-rc3-23925-00",
     "System.Globalization.Extensions": "4.0.1-rc3-23925-00",
     "System.IO": "4.1.0-rc3-23925-00",
+    "System.IO.FileSystem": "4.0.1-rc3-23925-00",
     "System.Linq.Expressions": "4.0.11-rc3-23925-00",
     "System.ObjectModel": "4.0.12-rc3-23925-00",
     "System.Reflection": "4.1.0-rc3-23925-00",
